+++ app/Http/Controllers/backend/ProductController.php
@@ -14,7 +14,8 @@ use App\Model\ProductImage;
 use App\Http\Requests\StoreProductRequest;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\Storage;
-
+use Illuminate\Support\Facades\Gate;
+use Auth;
 class ProductController extends Controller
 {
     /**
@@ -27,19 +28,9 @@ class ProductController extends Controller
     // const NỮ::0;
     public function index()
     {
-        // Storage::disk('public')->put('test1.txt', 'hoan');
-        // $file = Storage::files('/');
-        // $file = Storage::allFiles();
-        // Storage::disk('public')->delete('abc/test.txt');
-        // Storage::deleteDirectory($path);
-        // Storage::makeDirectory('category');
-        // $conten = Storage::disk('test')->get('test.txt');
-        // return Storage::download('test1.txt');
-        // dd(1);
         $categories = Category::all();
         $collection = Collection::all();
-        $products = Product::Paginate(5);
-
+        $products = Product::orderBy('updated_at','desc')->Paginate(5);
         return view('backend.product.index',[
             'products' => $products,
             'categories' => $categories,
@@ -57,10 +48,8 @@ class ProductController extends Controller
         
         $categories = Category::all();
         $collection = Collection::all();
-        $users = User::all();
         return view('backend.product.create',[
             'categories' => $categories,
-            'users' => $users,
             'collection' => $collection,
         ]);
     }
@@ -112,21 +101,30 @@ class ProductController extends Controller
         // }
        
         $data = $request->all();
+        $type_avatar = strstr($data['avatar']->getClientOriginalName(), '.');
         $images = $data['images'];
+        $parent_file = $data['name'].'/';
+        $information_product = $data['information_product'];
+        $user = Auth::user();
+
         $product = new Product();
 
         $product->name = $data['name'];     
         $product->category_id = $data['category_id'];
-        $product->user_id = $data['user_id'];
+        $product->user_id = $user->id;
         $product->collection_id = $data['collection_id'];
         $product->gender = $data['gender'];
-        $product->price = $data['price'];
-        $product->discount_percent = $data['discount_percent'];
+        $product->content = htmlspecialchars($data['content']);
+        $product->information_product = json_encode($information_product);
+        $product->origin_price = $data['origin_price'];
+        $product->sale_price = $data['sale_price'];
         $product->amount = $data['amount'];
+        $product->avatar = $parent_file.$data['avatar']->getClientOriginalName();
+        Storage::disk('public')->putFileAs($parent_file,$data['avatar'],'avatar'.$type_avatar);
 
         $product->save();
+
         $product_id = $product->id;
-        $parent_file = $data['name'].'/';
         if (isset($images)) {
             foreach ($images as $image) {
                 $img = new ProductImage();
@@ -155,11 +153,10 @@ class ProductController extends Controller
     public function show($id)
     {
         $product = Product::find($id);
-
-        return view('backend.product.detail',[
+        return view('fondend.detail_product',[
             'product' => $product,
 
-        ]); 
+        ]);     
     }
 
     /**
@@ -170,16 +167,24 @@ class ProductController extends Controller
      */
     public function edit($id)
     {
-        $categories = Category::all();
-        $collection = Collection::all();
-        $users = User::all();
+       
         $product = Product::find($id);
-        return view('backend.product.edit',[
-            'product' => $product,
-            'categories' => $categories,
-            'users' => $users,
-            'collection' => $collection,
-        ]);
+        $user = Auth::user();
+
+        if ($user->can('update',$product)) {
+            $categories = Category::all();
+            $collection = Collection::all();
+            return view('backend.product.edit',[
+                'product' => $product,
+                'categories' => $categories,
+                'collection' => $collection,
+            ]);
+        }else {
+            echo 'bạn không phải người đăng sản phẩm';
+            // return redirect()->route('product.index');
+        }
+
+        
     }
 
     /**
@@ -189,40 +194,80 @@ class ProductController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(StoreProductRequest $request, $id)
+    public function update(Request $request, $id)
     {
         $data = $request->all();
 
+        
+        // $user = $product->user;
+
+        // dd($product);
+        // $user = Auth::user();
+        // if ($user->can('update',$product)) {
+        //     echo 'aaaaa';
+        // }else {
+        //     echo 'sdsss';
+        // }
+        // dd();
+
+        // if (Gate::allows('update-product', $product)) {
+        //     $product->name = $data['name'];     
+        //     $product->category_id = $data['category_id'];
+        //     $product->user_id = $data['user_id'];
+        //     $product->collection_id = $data['collection_id'];
+        //     $product->gender = $data['gender'];
+        //     $product->price = $data['price'];
+        //     $product->discount_percent = $data['discount_percent'];
+
+        //     $product->save();
+        //     return redirect()->route('product.index');
+        // }else {
+        //     echo 'ko thanh cong';
+        // }
+
+        $data = $request->all();
+        // $type_avatar = strstr($data['avatar']->getClientOriginalName(), '.');
+        // $images = $data['images'];
+        $parent_file = $data['name'].'/';
+        $information_product = $data['information_product'];
+        $user = Auth::user();
+
         $product = Product::find($id);
+
         $product->name = $data['name'];     
         $product->category_id = $data['category_id'];
-        $product->user_id = $data['user_id'];
         $product->collection_id = $data['collection_id'];
         $product->gender = $data['gender'];
-        $product->price = $data['price'];
-        $product->discount_percent = $data['discount_percent'];
+        // $product->content = htmlspecialchars($data['content']);
+        $product->information_product = json_encode($information_product);
+        $product->origin_price = $data['origin_price'];
+        $product->sale_price = $data['sale_price'];
+        $product->amount = $data['amount'];
+        // $product->avatar = $parent_file.$data['avatar']->getClientOriginalName();
+        // Storage::disk('public')->putFileAs($parent_file,$data['avatar'],'avatar'.$type_avatar);
 
         $product->save();
-        $product_id = $product->id;
-        if (isset($data['productImage'])) {
-            foreach ($data['productImage'] as $file) {
-                $image = new ProductImage();
-                if ($request->hasFile('productImage')) {
-                   $image->name = $data['name'];
-                   $image->path = $file->getClientOriginalName();
-                   $file->move('image',$file->getClientOriginalName());
-                   $image->product_id = $product_id;
 
-                   $image->save();
-                }else {
-                    echo 'fail';
-                }
-            }
-        }else {
-            
-        }
+        // $product_id = $product->id;
+        // if (isset($images)) {
+        //     foreach ($images as $image) {
+        //         $img = new ProductImage();
+        //         $nameFile = $image->getClientOriginalName();             
+        //         $img->name = $data['name'];
+        //         $img->path = $parent_file.$nameFile;
+        //         $img->product_id = $product_id;
+        //         Storage::disk('public')->putFileAs($parent_file,$image,$nameFile);
+
+        //         $img->save();
+        //     }
+        // }else {
+        //     echo 'fail';
+        // }
 
         return redirect()->route('product.index');
+
+
+       
     }
 
     /**
@@ -234,15 +279,22 @@ class ProductController extends Controller
     public function destroy($id)
     {
         $product = Product::find($id);
-        $name = $product->name;
-        Storage::disk('public')->deleteDirectory($name);
-        $images = ProductImage::all()->where('product_id',$id); 
-        foreach ($images as $image) {
-            $image->delete();
-        }
-        $product->delete();
+        $user = Auth::user();
+        
+        if ($user->can('delete',$product)) {
+            $name = $product->name;
+            Storage::disk('public')->deleteDirectory($name);
+            $images = ProductImage::all()->where('product_id',$id); 
+            foreach ($images as $image) {
+                $image->delete();
+            }
+            $product->delete();
 
-        return redirect()->route('product.index');
+            return redirect()->route('product.index');
+        }else {
+            echo 'bạn không phải người đăng sản phẩm '.'<br>';
+            echo 'không được xóa';
+        }
     }
     public function showImages($id){
             $showImages = Product::find($id)->images;
